make tfp optional and dev analysis perturbation
